requirements.txt
Django~=4.0.4
djangorestframework~=3.13.1
djangorestframework-simplejwt
django-cors-headers
psycopg2~=2.9.3
pandas~=1.4.2